restructure playgrounds, fix imports
