+++ aihub_agent/playground/minimal_workflow/llama_index_workflow/LlamaIndexAgent.py
@@ -2,9 +2,7 @@ from aihub_agent.agents.abstract.Agent import Agent
 from aihub_agent.displayers.EventDisplayer import EventDisplayer
 from aihub_agent.workflow.decorators.step import step
 from aihub_lib.nats.events import StartEvent, StopEvent, LLMEvent, UserMessageEvent
-from playground.minimal_workflow.LlamaIndexAgent.LlamaIndexAgentConfig import (
-    LlamaIndexAgentConfig,
-)
+from playground.minimal_workflow.llama_index_workflow.LlamaIndexAgentConfig import LlamaIndexAgentConfig
 
 
 class LlamaIndexAgent(Agent):
